finished restructuring and basic testing for functionality

+++ src/supabase/depreciated/teamPlayers.ts
@@ -1,38 +1,37 @@
 /* 
 
-* This file contains all the 'single players' table related functions
+* This file contains all the 'team players' table related functions
 *
 *
 
 */
 
-import { supabase } from "supabaseDB/index";
 import { throwFormattedErrorLog } from "utils/logging/errorFormats";
+import { supabase } from "supabaseDB/index";
 import { addUser, getUserId, getUsernameAndDiscordId } from "./users";
 
 interface addPlayer {
-  username: string;
   discordUserId: string;
-  eventId: number;
-  discordServerId?: string;
+  teamId: number;
+  discordServerId: string;
+  username: string;
 }
 
 interface removePlayer {
-  username?: string;
   discordUserId: string;
-  eventId: number;
+  teamId: number;
 }
 
-interface getAllPlayers {
-  eventId: number;
+interface getAllTeamPlayers {
+  teamId: number;
 }
 
-interface getNumOfPlayers {
-  eventId: number;
+interface getNumOfTeamPlayers {
+  teamId: number;
 }
 
 export const addPlayer = async (props: addPlayer) => {
-  const { username, discordUserId, eventId, discordServerId } = props;
+  const { discordUserId, teamId, discordServerId, username } = props;
 
   let dbUserId: number;
 
@@ -40,6 +39,8 @@ export const addPlayer = async (props: addPlayer) => {
   const { data: getUserIdData, error: getUserIdError } = await getUserId({
     discordUserId: discordUserId,
   });
+
+  if (getUserIdError) throw throwFormattedErrorLog(getUserIdError);
 
   // add user to the Supabase DB if the user does not exist
   if (getUserIdData!.length < 1) {
@@ -49,14 +50,16 @@ export const addPlayer = async (props: addPlayer) => {
       discordServerId: discordServerId!,
     });
 
+    if (addUserError) throw throwFormattedErrorLog(addUserError);
+
     dbUserId = addUserData![0]["id"];
   } else {
     dbUserId = getUserIdData![0]["id"];
   }
 
-  const { data, error } = await supabase.from("SinglePlayers").insert([
+  const { data, error } = await supabase.from("TeamPlayers").insert([
     {
-      eventId: eventId,
+      teamId,
       userId: dbUserId,
     },
   ]);
@@ -67,38 +70,32 @@ export const addPlayer = async (props: addPlayer) => {
 };
 
 export const removePlayer = async (props: removePlayer) => {
-  const { discordUserId, eventId } = props;
+  const { discordUserId, teamId } = props;
 
   // get user ID from DB
   const { data: getUserIdData, error: getUserIdError } = await getUserId({
     discordUserId: discordUserId,
   });
 
+  if (getUserIdError) throw throwFormattedErrorLog(getUserIdError);
+
   const { data, error } = await supabase
-    .from("SinglePlayers")
+    .from("TeamPlayers")
     .delete()
     .eq("userId", getUserIdData![0]["id"])
-    .eq("eventId", eventId);
+    .eq("teamId", teamId);
 
   if (error) throw throwFormattedErrorLog(error);
 
   return { data, error };
 };
 
-interface Player {
-  username: string;
-  userId: string;
-}
-
-export const getAllPlayers = async (eventId: number): Promise<Player[]> => {
-  const { data, error } = await supabase
-    .from("SinglePlayers")
-    .select("userId")
-    .eq("eventId", eventId);
+export const getAllTeamPlayers = async (teamId: number) => {
+  const { data, error } = await supabase.from("TeamPlayers").select("userId").eq("teamId", teamId);
 
   if (error) throw throwFormattedErrorLog(error);
 
-  const players: Player[] = [];
+  const players = [];
 
   if (data) {
     for (const { userId } of data) {
@@ -110,38 +107,10 @@ export const getAllPlayers = async (eventId: number): Promise<Player[]> => {
   return players;
 };
 
-// export const getAllPlayers = async (props: getAllPlayers) => {
-// 	const { eventId } = props;
+export const getNumOfTeamPlayers = async (props: getNumOfTeamPlayers) => {
+  const { teamId } = props;
 
-// 	const players: [{ username: string; userId: string }?] = [];
-
-// 	const { data, error } = await supabase
-// 		.from('SinglePlayers')
-// 		.select('userId')
-// 		.eq('eventId', eventId);
-
-// 	if (error) throw throwFormattedErrorLog(error);
-
-// 	if (data)
-// 		for (const d of data!) {
-// 			const us: [{ username: string }] | any = await getUsernameAndDiscordId(
-// 				{
-// 					userId: d.userId,
-// 				},
-// 			);
-// 			players.push({
-// 				username: us[0]['username'],
-// 				userId: us[0]['userId'].toString(),
-// 			});
-// 		}
-
-// 	return players;
-// };
-
-export const getNumOfPlayers = async (props: getNumOfPlayers) => {
-  const { eventId } = props;
-
-  const { data, error } = await supabase.from("SinglePlayers").select("id").eq("eventId", eventId);
+  const { data, error } = await supabase.from("TeamPlayers").select("id").eq("teamId", teamId);
 
   if (error) throw throwFormattedErrorLog(error);
